Move error interfaces into error.ts
